cleanup/warnings (#111)
* build: suppress or resolve warnings

removed unused variables/imports, surpressed executable stack warning for tests (executables), reduce temporary array creation, avoided implicit type conversion, use `not_same` instead of exact comparison for reals

+++ test/helpers/bc_file.f90
@@ -42,7 +42,7 @@ contains
                 n(idx) = i_n
             end do
         end do
-        n = n*field%nfp
+        n = n*nint(field%nfp)
 
         allocate (rmnc(number_of_modes))
         allocate (zmns(number_of_modes))
